+++ lib/req.ex
@@ -116,7 +116,8 @@ defmodule Req do
           :cache_dir,
           :plug,
           :finch,
-          :finch_options,
+          :receive_timeout,
+          :pool_timeout,
           :http2,
           :unix_socket
         ])
@@ -510,8 +511,9 @@ defmodule Req do
 
     * `:finch` - the Finch pool to use. Defaults to pool automatically started by `Req`.
 
-    * `:finch_options` - options passed down to Finch when making the request, defaults to `[]`.
-       See `Finch.request/3` for a list of available options.
+    * `:pool_timeout` - pool checkout timeout in milliseconds, defaults to `5000`.
+
+    * `:receive_timeout` - socket receive timeout in milliseconds, defaults to `15_000`.
 
     * `:http2` - if `true`, uses an HTTP/2 pool automatically started by Req.